更新 updateEdgePaddingDelta 方法以接受整数参数并转换为双精度，优化 ReaderPage 组件的布局代码

+++ lib/reader/ui/components/sheets/font.dart
@@ -143,7 +143,9 @@ class _FontSheetContentState extends ConsumerState<FontSheetContent> {
                       edgePaddingDelta.value = val;
                     },
                     onChangeEnd: (val) {
-                      ref.read(ProviderUsecase().config.notifier).updateEdgePaddingDelta(edgePaddingDelta.value);
+                      ref
+                          .read(ProviderUsecase().config.notifier)
+                          .updateEdgePaddingDelta(edgePaddingDelta.value.toInt());
                     }),
               )),
               SvgPicture.asset(
